Move common.GetContainerResourcesInStatefulSet to pkg/utils/kubernetes
