+++ pkg/operation/common/utils_test.go
@@ -30,8 +30,8 @@ import (
 	"github.com/gardener/gardener/pkg/utils"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/test"
-	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
 
+	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
@@ -43,7 +43,6 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	schedulingv1 "k8s.io/api/scheduling/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/component-base/version"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -499,107 +498,6 @@ var _ = Describe("common", func() {
 			}
 
 			Expect(CheckIfDeletionIsConfirmed(obj)).To(Succeed())
-		})
-	})
-
-	Describe("#GetContainerResourcesInStatefulSet", func() {
-		var (
-			ctrl              *gomock.Controller
-			c                 *mockclient.MockClient
-			testNamespace     string
-			testStatefulset   string
-			statefulSet       *appsv1.StatefulSet
-			expectedResources *corev1.ResourceRequirements
-		)
-
-		BeforeEach(func() {
-			expectedResources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("100m"),
-					corev1.ResourceMemory: resource.MustParse("300Mi"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("1"),
-					corev1.ResourceMemory: resource.MustParse("3000Mi"),
-				},
-			}
-
-			ctrl = gomock.NewController(GinkgoT())
-			c = mockclient.NewMockClient(ctrl)
-			testNamespace = "test-namespace"
-			testStatefulset = "test-loki"
-
-			statefulSet = &appsv1.StatefulSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      testStatefulset,
-					Namespace: testNamespace,
-				},
-				Spec: appsv1.StatefulSetSpec{
-					Template: corev1.PodTemplateSpec{
-						Spec: corev1.PodSpec{},
-					},
-				},
-			}
-		})
-
-		AfterEach(func() {
-			ctrl.Finish()
-		})
-
-		It("should return container resources when statefulset contains one container", func() {
-			var (
-				ctx = context.TODO()
-			)
-
-			statefulSet.Spec.Template.Spec.Containers = []corev1.Container{
-				{
-					Name:      "container-1",
-					Resources: *expectedResources,
-				},
-			}
-
-			c.EXPECT().Get(ctx, kutil.Key(testNamespace, testStatefulset), gomock.AssignableToTypeOf(&appsv1.StatefulSet{})).SetArg(2, *statefulSet).Return(nil)
-
-			rr, err := GetContainerResourcesInStatefulSet(ctx, c, kutil.Key(testNamespace, testStatefulset))
-			Expect(err).NotTo(HaveOccurred())
-			Expect(rr).To(HaveLen(len(statefulSet.Spec.Template.Spec.Containers)))
-			Expect(rr["container-1"]).To(Equal(expectedResources))
-		})
-
-		It("should return all container resources when statefulset contains two containers", func() {
-			var (
-				ctx = context.TODO()
-			)
-
-			statefulSet.Spec.Template.Spec.Containers = []corev1.Container{
-				{
-					Name:      "container-1",
-					Resources: *expectedResources,
-				},
-				{
-					Name:      "container-2",
-					Resources: *expectedResources,
-				},
-			}
-
-			c.EXPECT().Get(ctx, kutil.Key(testNamespace, testStatefulset), gomock.AssignableToTypeOf(&appsv1.StatefulSet{})).SetArg(2, *statefulSet).Return(nil)
-
-			rr, err := GetContainerResourcesInStatefulSet(ctx, c, kutil.Key(testNamespace, testStatefulset))
-			Expect(err).NotTo(HaveOccurred())
-			Expect(rr).To(HaveLen(len(statefulSet.Spec.Template.Spec.Containers)))
-			Expect(rr["container-1"]).To(Equal(expectedResources))
-			Expect(rr["container-2"]).To(Equal(expectedResources))
-		})
-
-		It("should return error if statefulSet is not found", func() {
-			var (
-				ctx = context.TODO()
-			)
-
-			c.EXPECT().Get(ctx, kutil.Key(testNamespace, testStatefulset), gomock.AssignableToTypeOf(&appsv1.StatefulSet{})).Return(errors.New("error"))
-
-			_, err := GetContainerResourcesInStatefulSet(ctx, c, kutil.Key(testNamespace, testStatefulset))
-			Expect(err).To(HaveOccurred())
 		})
 	})
