+++ pkg/utils/kubernetes/statefulset.go
@@ -15,9 +15,13 @@
 package kubernetes
 
 import (
+	"context"
+
 	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // StatefulSetSource is a function that produces a slice of StatefulSets or an error.
@@ -99,4 +103,23 @@ func (d *statefulSetNamespaceLister) List(selector labels.Selector) ([]*appsv1.S
 	return filterStatefulSets(d.source, func(statefulSet *appsv1.StatefulSet) bool {
 		return statefulSet.Namespace == d.namespace && selector.Matches(labels.Set(statefulSet.Labels))
 	})
+}
+
+// GetContainerResourcesInStatefulSet returns the containers resources in StatefulSet.
+func GetContainerResourcesInStatefulSet(ctx context.Context, k8sClient client.Client, key client.ObjectKey) (map[string]*corev1.ResourceRequirements, error) {
+	statefulSet := &appsv1.StatefulSet{}
+	if err := k8sClient.Get(ctx, key, statefulSet); err != nil {
+		if apierrors.IsNotFound(err) {
+			return nil, nil
+		}
+		return nil, err
+	}
+
+	resourcesPerContainer := make(map[string]*corev1.ResourceRequirements)
+
+	for _, container := range statefulSet.Spec.Template.Spec.Containers {
+		resourcesPerContainer[container.Name] = container.Resources.DeepCopy()
+	}
+
+	return resourcesPerContainer, nil
 }
